requirements.txt
revChatGPT~=4.1.1.1
python-telegram-bot~=20.1
